requirements.txt
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
blinker==1.9.0
boto3==1.34.162
botocore==1.34.162
celery==5.5.3
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.43.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
redis==5.0.4
reportlab==4.0.4
requests==2.32.4
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stripe==12.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3

# === Añadidos necesarios para producción ===
gunicorn==22.0.0
psycopg2-binary==2.9.9
